+++ src/components/thumbnail-maker/types.d.ts
@@ -1,0 +1,1 @@
+export type Tag = { text: string; tagVariant: TagVariant; tagShape: TagShape };
